# docs/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a HR management system API.",
        "title": "HR Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/leave/leave-requests": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new leave request for approval",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leave Requests"
                ],
                "summary": "Submit a new leave request",
                "parameters": [
                    {
                        "description": "Leave Request object",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_leave_model.LeaveRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_leave_model.LeaveRequest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/leave/leave-requests/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a list of leave requests submitted by the current employee",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leave Requests"
                ],
                "summary": "List own leave requests",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Employee ID (simulation of auth context)",
                        "name": "employeeId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_leave_model.LeaveRequest"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/leave/leave-requests/pending-approvals": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a list of leave requests pending approval for a specific manager",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leave Requests"
                ],
                "summary": "List pending leave requests for approval",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Manager ID (simulation of auth context)",
                        "name": "managerId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_leave_model.LeaveRequest"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/leave/leave-requests/{id}/approve": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Approve a pending leave request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leave Requests"
                ],
                "summary": "Approve a leave request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Leave Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Manager ID (simulation of auth context)",
                        "name": "managerId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Approval note",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/leave/leave-requests/{id}/reject": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Reject a pending leave request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leave Requests"
                ],
                "summary": "Reject a leave request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Leave Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Manager ID (simulation of auth context)",
                        "name": "managerId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Rejection note",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_thecemakin_hr-project_internal_modules_corehr_model.Department": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "e.g., \"ENG\", \"MKT\", \"FIN\"",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "head": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Employee"
                        }
                    ]
                },
                "head_id": {
                    "description": "ID of the department head (Employee)",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_thecemakin_hr-project_internal_modules_corehr_model.Employee": {
            "type": "object",
            "properties": {
                "address_line1": {
                    "description": "Address Information",
                    "type": "string"
                },
                "address_line2": {
                    "type": "string"
                },
                "bank_account_number": {
                    "description": "Bank Account Details",
                    "type": "string"
                },
                "bank_name": {
                    "type": "string"
                },
                "bank_routing_number": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "department": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Department"
                        }
                    ]
                },
                "department_id": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "emergency_contact_name": {
                    "description": "Emergency Contact",
                    "type": "string"
                },
                "emergency_contact_phone": {
                    "type": "string"
                },
                "emergency_contact_relation": {
                    "type": "string"
                },
                "employee_number": {
                    "description": "Employment Information",
                    "type": "string"
                },
                "first_name": {
                    "description": "Personal Information",
                    "type": "string"
                },
                "gender": {
                    "description": "male, female, other",
                    "type": "string"
                },
                "hire_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "manager": {
                    "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Employee"
                },
                "manager_id": {
                    "description": "Self-referencing foreign key",
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "position": {
                    "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Position"
                },
                "position_id": {
                    "type": "integer"
                },
                "postal_code": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "status": {
                    "description": "active, inactive, terminated",
                    "type": "string"
                },
                "subordinates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Employee"
                    }
                },
                "termination_date": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_thecemakin_hr-project_internal_modules_corehr_model.Position": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "e.g., \"SENG\", \"MGR\", \"VP\"",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "department": {
                    "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Department"
                },
                "department_id": {
                    "description": "Relationships",
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "level": {
                    "description": "Hierarchical level (1-10 scale)",
                    "type": "integer"
                },
                "salary_max": {
                    "description": "Maximum salary in cents",
                    "type": "integer"
                },
                "salary_min": {
                    "description": "Minimum salary in cents",
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_thecemakin_hr-project_internal_modules_leave_model.LeaveRequest": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "employee": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Employee"
                        }
                    ]
                },
                "employee_id": {
                    "type": "integer"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "leave_type": {
                    "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_leave_model.LeaveType"
                },
                "leave_type_id": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "reviewer": {
                    "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Employee"
                },
                "reviewer_id": {
                    "type": "integer"
                },
                "reviewer_note": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "description": "pending, approved, rejected",
                    "type": "string"
                },
                "total_days_requested": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_thecemakin_hr-project_internal_modules_leave_model.LeaveType": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "default_days": {
                    "type": "integer"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer \u003cyour-jwt-token\u003e\" to authenticate.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}